refactor(doctor): extract telegram allowfrom repair
